Add all scenarios in the test case
